Fixed call to pow for windows/VS.
git-svn-id: http://svn.code.sf.net/p/protomol/svn/trunk@1340 ce46b6ef-3445-0410-911d-c47028e6f9df

+++ protomol/protomol/integrator/hessian/HessianInt.cpp
@@ -133,7 +133,7 @@ void HessianInt::initialize(ProtoMolApp *app) {
   //automatically generate parameters?
   if(autoParmeters){
     numberOfModes = 3*(int)sqrt((float)sz);
-    residuesPerBlock = (int)pow((float)_N,0.6) / 15;
+    residuesPerBlock = (int)pow((double)_N,0.6) / 15;
     blockVectorCols = 10 + (int)sqrt((float)residuesPerBlock);
     blockCutoffDistance = hsn.cutOff;  
     report << hint << "[HessianInt::initialize] Auto parameters: numberOfModes " << numberOfModes <<
